% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{minifyr_html_minifier}
\alias{minifyr_html_minifier}
\title{Minify HTML}
\usage{
minifyr_html_minifier(input, output)
}
\arguments{
\item{input}{File to minify}

\item{output}{Name of the output file}
}
\value{
The output path
}
\description{
Minify HTML files.
}
\details{
Available minifiers are:
\itemize{
\item \href{https://github.com/kangax/html-minifier}{html-minifier}, with \code{minifyr_html_minifier()}
}
}
