% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{minifyr_json_jsonminify}
\alias{minifyr_json_jsonminify}
\title{Minify JSON Files}
\usage{
minifyr_json_jsonminify(input, output)
}
\arguments{
\item{input}{File to minify}

\item{output}{Name of the output file}
}
\value{
The output path
}
\description{
Minify JSON files.
}
\details{
Available minifiers are:
\itemize{
\item \href{https://github.com/fkei/JSON.minify}{jsonminify}, with \code{minifyr_json_jsonminify()}
}
}
