% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{minify_folder_css}
\alias{minify_folder_css}
\alias{minify_package_css}
\alias{minify_folder_js}
\alias{minify_package_js}
\title{Minifying folders}
\usage{
minify_folder_css(path, fun, recursive = TRUE, ...)

minify_package_css(package, fun, recursive = TRUE, ...)

minify_folder_js(path, fun, recursive = TRUE, ...)

minify_package_js(package, fun, recursive = TRUE, ...)
}
\arguments{
\item{path}{Path to the folder containing files to minify}

\item{fun}{Function to use to minify}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{...}{Parameters passed to \code{base::list.files()}}

\item{package}{Package containing files to minify}
}
\value{

}
\description{
Minifying folders
}
