% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{minifyr_example}
\alias{minifyr_example}
\title{Return the path to a file example}
\usage{
minifyr_example(ext = c("html", "js", "json", "css"))
}
\arguments{
\item{ext}{Extension to use}
}
\value{
a path
}
\description{
Return the path to a file example
}
\examples{
minifyr_example("css")
}
