% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{minifyr_npm_install}
\alias{minifyr_npm_install}
\title{Run NPM install}
\usage{
minifyr_npm_install(force = FALSE)
}
\arguments{
\item{force}{should the user be prompted to install?}
}
\value{
An installed lib
}
\description{
Run NPM install
}
