% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{minifyr_run}
\alias{minifyr_run}
\title{Title}
\usage{
minifyr_run(input, output)
}
\arguments{
\item{input}{CSS file to minify}

\item{output}{Name of the output file}
}
\value{
The output path
}
\description{
Title
}
