% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_is_css}
\alias{check_is_css}
\alias{check_is_html}
\alias{check_is_js}
\alias{check_is_json}
\title{Verify the extension of a file}
\usage{
check_is_css(file)

check_is_html(file)

check_is_js(file)

check_is_json(file)
}
\arguments{
\item{file}{path to a file}
}
\value{

}
\description{
Verify the extension of a file
}
\examples{
check_is_css("path/to/css")
}
